Move iterator to a separate file
